LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/srcdoc/consecutive-srcdoc-expected.txt


FAIL changing srcdoc does a replace navigation since the URL is still about:srcdoc assert_equals: expected 2 but got 1
FAIL changing srcdoc to about:srcdoc#yo then another srcdoc does two push navigations and we can navigate back assert_equals: expected 2 but got 1
